Fixed issues related to incorporating Ignite into existing project via "add package": * endless loop scanning path looking for Packages.swift * permission problems with deleting and recreating Build directory unless it is in the Sandbox's container
